sheguard-backend/requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
filelock==3.17.0
Flask==2.3.2
Flask-Cors==5.0.1
fsspec==2025.2.0
gunicorn==23.0.0
huggingface-hub==0.29.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.23.5
opencv-python==4.8.1.78
packaging==24.2
Pillow==10.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sympy==1.13.3
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.67.1
transformers==4.31.0
typing_extensions==4.12.2
urllib3==2.3.0
waitress==3.0.2
Werkzeug==3.1.3
setuptools>=78.0.2
wheel
scikit-learn==1.3.0
scipy==1.11.1